Fix resizing issue
WebKitWebView always asks for enough size to display its contents. So,
when you expand the window the content size gets bigger, but then you're
not be able to shrink it anymore.

To fix this issue, simply use the set_size_request() method to set the
minimum size of the webview.
The size request is set to 800x600 which is the smallest size CryptoCat
can accept while still functioning well and drawing itself correctly.

Also, this will always make the window to be placed in the center of the
screen.